Move jwt to middleware
